==> Code/01_Data_Wrangling_and_Cleaning.ipynb <==
# <u>Data Wrangling and Cleaning
---------

## Objective
The goal of this first notebook is to combine and clean separate each of the 50 players' data. This includes dropping unnecessary columns, unfinished golf tournaments, and creating a Strokes Gained: Total column. The result is a single dataframe of 4,648 golf tournaments and 83 features.

------
#### External Libraries Import

import pandas as pd
import numpy as np
from functions import fetch  # functions.py file
from os import listdir
import warnings
warnings.filterwarnings('ignore')

#### Fetch and Observe Data for 50 players

# create list of all players in the Data folder
players = listdir('../Data/')

# remove unwanted files/folders
players.remove('.DS_Store') 
players.remove('.ipynb_checkpoints')
players.remove('Sets')
# print length of list
print(f'{len(players)} players')

# fetch data for all 50 players using the fetch function in .py file
df_pga = fetch(players)

# observe raw data
df_pga.head()

- Each observation in the dataframe is one players' stats for a single golf tournament. Player name is under the 'player' column on the far left.

### <u>Drop unwanted rows and columns

# replace unfinished tournaments with NaNs
df_pga.replace('CUT', np.nan, inplace = True)
df_pga.replace('W/D', np.nan, inplace = True)
df_pga.replace('DQ', np.nan, inplace = True)

- Only complete 4 round golf tournaments are used in this study.

# drop any row with missing data in any strokes gained column
df_pga.dropna(subset = ['sg:_off-the-tee', 
                        'sg:_approach-the-green', 
                        'sg:_around-the-green', 
                        'sg:_putting', 
                        'club_head_speed',
                        'finish'], inplace = True)
# drop irrelevant features
df_pga.drop(columns = ['finish_pos._x', 'finish_pos._y', 'tournament_x', 'tournament_y',
                       'sg:_tee-to-green', 'total_driving_efficiency',
                       'total_driving', 'total_putting', 'driving_distance_-_all_drives',
                       'putts_per_round_-_round_1','putts_per_round_-_round_2', 
                       'putts_per_round_-_round_3','putts_per_round_-_round_4', 
                       'average_distance_of_putts_made'], 
            inplace = True)

# check number of rows and columns
print(f'The dataset contains {df_pga.shape[0]} golf tournaments.')
print(f'The dataset includes {df_pga.shape[1]} golf metrics.')

- In the scope of this study, any feature that takes into account a "ranking" or "number of rounds played" does not matter. This analysis focuses on raw golf metrics that a golfer at any level of skill can measure.

### <u>Create a Strokes Gained: Total column

# list of each Strokes Gained category
sg_stats = ['sg:_off-the-tee', 
            'sg:_approach-the-green', 
            'sg:_around-the-green', 
            'sg:_putting']

# convert datatype to float
for stat in sg_stats:
    df_pga[stat] = df_pga[stat].astype(float)
    
# sum of off-the-tee, approach-the-green, around-the-green, and putting equals total
df_pga['sg:_total'] = df_pga['sg:_off-the-tee'] \
+ df_pga['sg:_approach-the-green'] \
+ df_pga['sg:_around-the-green'] \
+ df_pga['sg:_putting']

#### Save full dataset

df_pga.to_csv('../Data/Sets/full_pga_data.csv', index = False)
